# src/resources/npm/api-client/src/models/ServiceBody.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMLT - OpenAPI 3.0
 * BMLT Admin API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ServiceBody
 */
export interface ServiceBody {
    /**
     * 
     * @type {number}
     * @memberof ServiceBody
     */
    parentId?: number;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    type?: string;
    /**
     * 
     * @type {number}
     * @memberof ServiceBody
     */
    adminUserId?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof ServiceBody
     */
    assignedUserIds?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    helpline?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceBody
     */
    worldId?: string;
    /**
     * 
     * @type {number}
     * @memberof ServiceBody
     */
    id?: number;
}

/**
 * Check if a given object implements the ServiceBody interface.
 */
export function instanceOfServiceBody(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function ServiceBodyFromJSON(json: any): ServiceBody {
    return ServiceBodyFromJSONTyped(json, false);
}

export function ServiceBodyFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceBody {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'parentId': !exists(json, 'parentId') ? undefined : json['parentId'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'type': !exists(json, 'type') ? undefined : json['type'],
        'adminUserId': !exists(json, 'adminUserId') ? undefined : json['adminUserId'],
        'assignedUserIds': !exists(json, 'assignedUserIds') ? undefined : json['assignedUserIds'],
        'url': !exists(json, 'url') ? undefined : json['url'],
        'helpline': !exists(json, 'helpline') ? undefined : json['helpline'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'worldId': !exists(json, 'worldId') ? undefined : json['worldId'],
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function ServiceBodyToJSON(value?: ServiceBody | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'parentId': value.parentId,
        'name': value.name,
        'description': value.description,
        'type': value.type,
        'adminUserId': value.adminUserId,
        'assignedUserIds': value.assignedUserIds,
        'url': value.url,
        'helpline': value.helpline,
        'email': value.email,
        'worldId': value.worldId,
        'id': value.id,
    };
}