# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BMLT - OpenAPI 3.0",
        "description": "BMLT Admin API Documentation",
        "license": {
            "name": "MIT",
            "url": "https://github.com/bmlt-enabled/bmlt-root-server/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/servicebodies": {
            "get": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Retrieves service bodies",
                "description": "Retrieve service bodies for authenticated user.",
                "operationId": "getServiceBodies",
                "responses": {
                    "200": {
                        "description": "Returns when user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceBodyCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Creates a service body",
                "description": "Creates a service body.",
                "operationId": "createServiceBody",
                "requestBody": {
                    "description": "Pass in service body object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceBodyCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns when POST is successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceBody"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no service body exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoServiceBodyExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
        "/api/v1/servicebodies/{serviceBodyId}": {
            "get": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Retrieves a service body",
                "description": "Retrieve a single service body by id.",
                "operationId": "getServiceBody",
                "parameters": [
                    {
                        "name": "serviceBodyId",
                        "in": "path",
                        "description": "ID of service body",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns when user is authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceBody"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no service body exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoServiceBodyExists"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Updates a Service Body",
                "description": "Updates a single service body.",
                "operationId": "updateServiceBody",
                "parameters": [
                    {
                        "name": "serviceBodyId",
                        "in": "path",
                        "description": "ID of service body",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass in service body object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceBodyUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no service body exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoServiceBodyExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Deletes a service body",
                "description": "Deletes a service body by id.",
                "operationId": "deleteServiceBody",
                "parameters": [
                    {
                        "name": "serviceBodyId",
                        "in": "path",
                        "description": "ID of service body",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no service body exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoServiceBodyExists"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "rootServer"
                ],
                "summary": "Patches a service body",
                "description": "Patches a single service body by id.",
                "operationId": "patchServiceBody",
                "parameters": [
                    {
                        "name": "serviceBodyId",
                        "in": "path",
                        "description": "ID of service body",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass in fields you want to update.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceBodyPartialUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "401": {
                        "description": "Returns when user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Returns when user is unauthorized to perform action.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no service body exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoServiceBodyExists"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bmltToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorUnauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ErrorUnauthorized": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The field is required. (and 1 more error)"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "error details"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "ServiceBodyBase": {
                "properties": {
                    "parentId": {
                        "type": "integer",
                        "example": "0"
                    },
                    "name": {
                        "type": "string",
                        "example": "string"
                    },
                    "description": {
                        "type": "string",
                        "example": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "string"
                    },
                    "adminUserId": {
                        "type": "integer",
                        "example": "0"
                    },
                    "assignedUserIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": "0"
                        }
                    },
                    "url": {
                        "type": "string",
                        "example": "string"
                    },
                    "helpline": {
                        "type": "string",
                        "example": "string"
                    },
                    "email": {
                        "type": "string",
                        "example": "string"
                    },
                    "worldId": {
                        "type": "string",
                        "example": "string"
                    }
                },
                "type": "object"
            },
            "ServiceBody": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ServiceBodyBase"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "0"
                            }
                        }
                    }
                ]
            },
            "ServiceBodyCreate": {
                "required": [
                    "parentId",
                    "name",
                    "description",
                    "type",
                    "adminUserId",
                    "assignedUserIds"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ServiceBodyBase"
                    }
                ]
            },
            "ServiceBodyUpdate": {
                "required": [
                    "parentId",
                    "name",
                    "description",
                    "type",
                    "adminUserId",
                    "assignedUserIds"
                ],
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ServiceBodyBase"
                    }
                ]
            },
            "ServiceBodyPartialUpdate": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ServiceBodyBase"
                    }
                ]
            },
            "ServiceBodyCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/ServiceBody"
                }
            },
            "NoServiceBodyExists": {
                "description": "Returns when no user exists.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bmltToken": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "api/v1/auth/token",
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {}
                    }
                }
            }
        }
    }
}